+++ src/main/java/net/meander/subtlyd/world/item/CreativeModeTabsSD.java
@@ -50,6 +50,7 @@ public class CreativeModeTabsSD {
             entries.insertAfter(BUSH, REEDS);
             entries.insertAfter(WARPED_WART_BLOCK, WARPED_OVERHANG);
             entries.insertAfter(BASALT, BASALT_SLAB);
+            entries.insertAfter(JACK_O_LANTERN, SOUL_JACK_O_LANTERN);
         });
 
         CreativeModeTabEvents.modifyOutputEvent(CreativeModeTabs.COLORED_BLOCKS).register(entries -> {