+++ src/main/java/net/meander/subtlyd/world/item/ItemsSD.java
@@ -1,18 +1,14 @@
 package net.meander.subtlyd.world.item;
 
 import net.fabricmc.fabric.api.registry.CompostableRegistry;
+import net.meander.subtlyd.references.BlockItemIdsSD;
 import net.meander.subtlyd.references.ItemIdsSD;
-import net.meander.subtlyd.util.Util;
 import net.meander.subtlyd.world.block.BlocksSD;
 import net.meander.subtlyd.world.entity.EntityTypesSD;
 import net.meander.subtlyd.world.food.FoodsSD;
 import net.meander.subtlyd.world.item.component.ConsumablesSD;
-import net.minecraft.core.Registry;
 import net.minecraft.core.component.DataComponents;
-import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.core.registries.Registries;
 import net.minecraft.references.BlockItemId;
-import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
@@ -25,79 +21,58 @@ import net.minecraft.world.level.block.CampfireBlock;
 import net.minecraft.world.level.block.ColorCollection;
 
 import java.util.function.BiFunction;
-import java.util.function.Function;
 import java.util.function.UnaryOperator;
 
 public class ItemsSD {
-    public static final Item APPLE_PIE = registerItemSD(resourceKey("apple_pie"), Item::new, new Item.Properties().food(FoodsSD.APPLE_PIE));
-    public static final Item CALAMARI = registerItemSD(resourceKey("calamari"), Item::new, new Item.Properties().food(FoodsSD.CALAMARI));
-    public static final Item COOKED_CALAMARI = registerItemSD(resourceKey("cooked_calamari"), Item::new, new Item.Properties().food(FoodsSD.COOKED_CALAMARI));
-    public static final Item POTTAGE = registerItemSD(resourceKey("pottage"), Item::new, new Item.Properties().food(FoodsSD.POTTAGE).stacksTo(1));
-    public static final ColorCollection<Item> TENT = ColorCollection.registerItems(
-            ItemIdsSD.TENT, (id, color) -> registerItemSD(id, properties -> new TentItem(EntityTypesSD.TENT.pick(color), properties), new Item.Properties().stacksTo(1))
+    public static final Item APPLE_PIE = Items.registerItem(ItemIdsSD.APPLE_PIE, Item::new, new Item.Properties().food(FoodsSD.APPLE_PIE));
+    public static final Item CALAMARI = Items.registerItem(ItemIdsSD.CALAMARI, Item::new, new Item.Properties().food(FoodsSD.CALAMARI));
+    public static final Item COOKED_CALAMARI = Items.registerItem(ItemIdsSD.COOKED_CALAMARI, Item::new, new Item.Properties().food(FoodsSD.COOKED_CALAMARI));
+    public static final Item POTTAGE = Items.registerItem(ItemIdsSD.POTTAGE, Item::new, new Item.Properties().food(FoodsSD.POTTAGE).stacksTo(1));
+    public static final ColorCollection<Item> TENT = ColorCollection.registerItems(ItemIdsSD.TENT,
+            (id, color) -> Items.registerItem(id, properties -> new TentItem(EntityTypesSD.TENT.pick(color), properties), new Item.Properties().stacksTo(1))
     );
-    public static final Item UNLIT_CAMPFIRE = registerBlockSD(Blocks.CAMPFIRE, (properties -> properties.component(DataComponents.CONTAINER, ItemContainerContents.EMPTY).component(DataComponents.BLOCK_STATE, BlockItemStateProperties.EMPTY.with(CampfireBlock.LIT, false))), "unlit_campfire");
-    public static final Item SNOW_BRICKS = registerBlockSD(BlocksSD.SNOW_BRICKS);
-    public static final Item SNOW_BRICK_STAIRS = registerBlockSD(BlocksSD.SNOW_BRICK_STAIRS);
-    public static final Item SNOW_BRICK_SLAB = registerBlockSD(BlocksSD.SNOW_BRICK_SLAB);
-    public static final Item SNOW_BRICK_WALL = registerBlockSD(BlocksSD.SNOW_BRICK_WALL);
-    public static final Item CHARCOAL_BLOCK = registerBlockSD(BlocksSD.CHARCOAL_BLOCK);
-    public static final Item IRON_GRATE = registerBlockSD(BlocksSD.IRON_GRATE);
-    public static final Item CHISELED_POLISHED_DRIPSTONE = registerBlockSD(BlocksSD.CHISELED_POLISHED_DRIPSTONE);
-    public static final Item POLISHED_DRIPSTONE = registerBlockSD(BlocksSD.POLISHED_DRIPSTONE);
-    public static final Item POLISHED_DRIPSTONE_SLAB = registerBlockSD(BlocksSD.POLISHED_DRIPSTONE_SLAB);
-    public static final Item POLISHED_DRIPSTONE_STAIRS = registerBlockSD(BlocksSD.POLISHED_DRIPSTONE_STAIRS);
-    public static final Item POLISHED_DRIPSTONE_WALL = registerBlockSD(BlocksSD.POLISHED_DRIPSTONE_WALL);
-    public static final Item STONE_PILLAR = registerBlockSD(BlocksSD.STONE_PILLAR);
-    public static final Item STONE_TILES = registerBlockSD(BlocksSD.STONE_TILES);
-    public static final Item STONE_TILE_STAIRS = registerBlockSD(BlocksSD.STONE_TILE_STAIRS);
-    public static final Item STONE_TILE_SLAB = registerBlockSD(BlocksSD.STONE_TILE_SLAB);
-    public static final Item STONE_TILE_WALL = registerBlockSD(BlocksSD.STONE_TILE_WALL);
-    public static final Item REEDS = registerBlockSD(BlocksSD.REEDS);
-    public static final Item WARPED_OVERHANG = registerBlockSD(BlocksSD.WARPED_OVERHANG);
-    public static final Item BLAST_FUNGUS = registerItemSD("blast_fungus", BlastFungusItem::new, new Item.Properties().stacksTo(16));
-    public static final Item BASALT_SLAB = registerBlockSD(BlocksSD.BASALT_SLAB);
-    public static final Item ELIXIR = registerItemSD("elixir", Item::new, new Item.Properties().stacksTo(16).rarity(Rarity.UNCOMMON).component(DataComponents.CONSUMABLE, ConsumablesSD.ELIXIR).usingConvertsTo(Items.GLASS_BOTTLE));
+    public static final Item UNLIT_CAMPFIRE = registerBlockSD(BlockItemIdsSD.UNLIT_CAMPFIRE, Blocks.CAMPFIRE, (properties -> properties.component(DataComponents.CONTAINER, ItemContainerContents.EMPTY)
+            .component(DataComponents.BLOCK_STATE, BlockItemStateProperties.EMPTY.with(CampfireBlock.LIT, false))));
+    public static final Item SNOW_BRICKS = Items.registerBlock(BlockItemIdsSD.SNOW_BRICKS, BlocksSD.SNOW_BRICKS);
+    public static final Item SNOW_BRICK_STAIRS = Items.registerBlock(BlockItemIdsSD.SNOW_BRICK_STAIRS, BlocksSD.SNOW_BRICK_STAIRS);
+    public static final Item SNOW_BRICK_SLAB = Items.registerBlock(BlockItemIdsSD.SNOW_BRICK_SLAB, BlocksSD.SNOW_BRICK_SLAB);
+    public static final Item SNOW_BRICK_WALL = Items.registerBlock(BlockItemIdsSD.SNOW_BRICK_WALL, BlocksSD.SNOW_BRICK_WALL);
+    public static final Item CHARCOAL_BLOCK = Items.registerBlock(BlockItemIdsSD.CHARCOAL_BLOCK, BlocksSD.CHARCOAL_BLOCK);
+    public static final Item IRON_GRATE = Items.registerBlock(BlockItemIdsSD.IRON_GRATE, BlocksSD.IRON_GRATE);
+    public static final Item CHISELED_POLISHED_DRIPSTONE = Items.registerBlock(BlockItemIdsSD.CHISELED_POLISHED_DRIPSTONE, BlocksSD.CHISELED_POLISHED_DRIPSTONE);
+    public static final Item POLISHED_DRIPSTONE = Items.registerBlock(BlockItemIdsSD.POLISHED_DRIPSTONE, BlocksSD.POLISHED_DRIPSTONE);
+    public static final Item POLISHED_DRIPSTONE_SLAB = Items.registerBlock(BlockItemIdsSD.POLISHED_DRIPSTONE_SLAB, BlocksSD.POLISHED_DRIPSTONE_SLAB);
+    public static final Item POLISHED_DRIPSTONE_STAIRS = Items.registerBlock(BlockItemIdsSD.POLISHED_DRIPSTONE_STAIRS, BlocksSD.POLISHED_DRIPSTONE_STAIRS);
+    public static final Item POLISHED_DRIPSTONE_WALL = Items.registerBlock(BlockItemIdsSD.POLISHED_DRIPSTONE_WALL, BlocksSD.POLISHED_DRIPSTONE_WALL);
+    public static final Item STONE_PILLAR = Items.registerBlock(BlockItemIdsSD.STONE_PILLAR, BlocksSD.STONE_PILLAR);
+    public static final Item STONE_TILES = Items.registerBlock(BlockItemIdsSD.STONE_TILES, BlocksSD.STONE_TILES);
+    public static final Item STONE_TILE_STAIRS = Items.registerBlock(BlockItemIdsSD.STONE_TILE_STAIRS, BlocksSD.STONE_TILE_STAIRS);
+    public static final Item STONE_TILE_SLAB = Items.registerBlock(BlockItemIdsSD.STONE_TILE_SLAB, BlocksSD.STONE_TILE_SLAB);
+    public static final Item STONE_TILE_WALL = Items.registerBlock(BlockItemIdsSD.STONE_TILE_WALL, BlocksSD.STONE_TILE_WALL);
+    public static final Item REEDS = Items.registerBlock(BlockItemIdsSD.REEDS, BlocksSD.REEDS);
+    public static final Item WARPED_OVERHANG = Items.registerBlock(BlockItemIdsSD.WARPED_OVERHANG, BlocksSD.WARPED_OVERHANG);
+    public static final Item BLAST_FUNGUS = Items.registerItem(ItemIdsSD.BLAST_FUNGUS, BlastFungusItem::new, new Item.Properties().stacksTo(16));
+    public static final Item BASALT_SLAB = Items.registerBlock(BlockItemIdsSD.BASALT_SLAB, BlocksSD.BASALT_SLAB);
+    public static final Item ELIXIR = Items.registerItem(ItemIdsSD.ELIXIR, Item::new, new Item.Properties().stacksTo(16)
+            .rarity(Rarity.UNCOMMON)
+            .component(DataComponents.CONSUMABLE, ConsumablesSD.ELIXIR)
+            .usingConvertsTo(Items.GLASS_BOTTLE));
+    public static final Item SOUL_JACK_O_LANTERN = Items.registerBlock(BlockItemIdsSD.SOUL_JACK_O_LANTERN, BlocksSD.SOUL_JACK_O_LANTERN);
 
     public static void registration() {
         CreativeModeTabsSD.bootstrap();
         CompostableRegistry.INSTANCE.add(APPLE_PIE, 1.0F);
     }
 
-    private static Item registerItemSD(final String name, final Function<Item.Properties, Item> itemFactory, final Item.Properties properties) {
-        return registerItemSD(ResourceKey.create(Registries.ITEM, Util.identifier(name)), itemFactory, properties);
+    private static Item registerBlockSD(final BlockItemId id, final Block block, final UnaryOperator<Item.Properties> propertiesFunction) {
+        return registerBlockSD(id, block, (b, p) -> new BlockItem(b, propertiesFunction.apply(p)));
     }
 
-    public static Item registerItemSD(final ResourceKey<Item> key, final Function<Item.Properties, Item> itemFactory, final Item.Properties properties) {
-        Item item = itemFactory.apply(properties.setId(key));
-        return Registry.register(BuiltInRegistries.ITEM, key, item);
+    private static Item registerBlockSD(final BlockItemId id, final Block block, final BiFunction<Block, Item.Properties, Item> itemFactory) {
+        return registerBlockSD(id, block, itemFactory, new Item.Properties());
     }
 
-    private static Item registerBlockSD(Block block) {
-        return Items.registerBlock(BlockItemId.create(BuiltInRegistries.BLOCK.getKey(block), BuiltInRegistries.BLOCK.getKey(block)), block);
-    }
-
-    private static Item registerBlockSD(Block block, UnaryOperator<Item.Properties> unaryOperator, String location) {
-        return registerBlockSD(
-                block, ((blockx, properties) -> new BlockItem(blockx, unaryOperator.apply(properties))), location
-        );
-    }
-
-    private static Item registerBlockSD(Block block, BiFunction<Block, Item.Properties, Item> biFunction, String location) {
-        return registerBlockSD(block, biFunction, new Item.Properties(), location);
-    }
-
-    private static Item registerBlockSD(Block block, BiFunction<Block, Item.Properties, Item> biFunction, Item.Properties properties, String location) {
-        return registerItemSD(
-                blockIdToItemIdSD(location), propertiesx -> biFunction.apply(block, propertiesx), properties.useBlockDescriptionPrefix()
-        );
-    }
-
-    private static ResourceKey<Item> blockIdToItemIdSD(String location) {
-        return ResourceKey.create(Registries.ITEM, Util.identifier(location));
-    }
-
-    private static ResourceKey<Item> resourceKey(String name) {
-        return ResourceKey.create(Registries.ITEM, Util.identifier(name));
+    private static Item registerBlockSD(final BlockItemId id, final Block block, final BiFunction<Block, Item.Properties, Item> itemFactory, final Item.Properties properties) {
+        return Items.registerItem(id.item(), p -> itemFactory.apply(block, p), properties.useBlockDescriptionPrefix().requiredFeatures(block.requiredFeatures()));
     }
 }